src/channel-layer/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-cloud-pubsub==2.23.0
requests==2.32.0
psycopg2-binary==2.9.9
